refactor : path를 구하는 로직을 path 도메인에게 위임하는 리팩토링
